Add basic support for XML mappings
"Basic" because we consider that all XML mapping files are unique
in the classpath. Even if an XML mapping file is right next to the
persistence.xml that refers to it, if there is another one with the same
name in the classpath, we will fail.

That's annoying for use cases where persistence.xml files are imported
from libraries and each persistence.xml uses its own mapping file with
the same, standard name: META-INF/orm.xml.

But that will have to be solved separately.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>